4-3-20 (a)
Added logging of invalid class names to dynamicConfig function just in case thare are some.

Added a new function to removed blacklisted items from an array of items. Not sure this is coded the fasted way.

Added a function that is probably not currently used: selectRandomObject, which includes the ability to check against blacklists.

Added or debuged functions to check for nearby groups.